@@ -1,9 +1,10 @@
 name: demo_app
 description: Flutter Widget from HTML demo app.
 version: 0.0.1
+publish_to: "none"
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   carousel_slider: ^4.0.0
