@@ -14,7 +14,7 @@ class WordpressScreen extends StatelessWidget {
     'The Mozilla Blog': 'blog.mozilla.org',
   };
 
-  const WordpressScreen({Key key}) : super(key: key);
+  const WordpressScreen({super.key});
 
   @override
   Widget build(BuildContext context) => Scaffold(
@@ -37,9 +37,9 @@ class WordpressScreen extends StatelessWidget {
 
 class PostsScreen extends StatelessWidget {
   final String domain;
-  final String title;
+  final String? title;
 
-  const PostsScreen(this.domain, {Key key, this.title}) : super(key: key);
+  const PostsScreen(this.domain, {super.key, this.title});
 
   @override
   Widget build(BuildContext _) => Scaffold(
@@ -56,42 +56,45 @@ class PostsScreen extends StatelessWidget {
 class _PostScreen extends StatelessWidget {
   final _Post post;
 
-  const _PostScreen(this.post, {Key key}) : super(key: key);
+  const _PostScreen(this.post);
 
   @override
-  Widget build(BuildContext context) => SelectionAreaScaffold(
-        appBar: AppBar(
-          title: HtmlWidget(post.title),
-          actions: [
-            IconButton(
-              icon: const Icon(Icons.open_in_browser),
-              onPressed: () => launchUrl(Uri.parse(post.link)),
-            ),
-            const PopupMenu(
-              toggleIsSelectable: true,
-            ),
-          ],
-        ),
-        body: Padding(
-          padding: const EdgeInsets.all(8),
-          child: SingleChildScrollView(
-            child: Column(
-              children: [
-                if (post.featuredMedia != null)
-                  AspectRatio(
-                    aspectRatio:
-                        post.featuredMedia.width / post.featuredMedia.height,
-                    child: Center(
-                      child: Image.network(post.featuredMedia.sourceUrl),
-                    ),
+  Widget build(BuildContext context) {
+    final featuredMedia = post.featuredMedia;
+
+    return SelectionAreaScaffold(
+      appBar: AppBar(
+        title: HtmlWidget(post.title),
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.open_in_browser),
+            onPressed: () => launchUrl(Uri.parse(post.link)),
+          ),
+          const PopupMenu(
+            toggleIsSelectable: true,
+          ),
+        ],
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(8),
+        child: SingleChildScrollView(
+          child: Column(
+            children: [
+              if (featuredMedia != null)
+                AspectRatio(
+                  aspectRatio: featuredMedia.width / featuredMedia.height,
+                  child: Center(
+                    child: Image.network(featuredMedia.sourceUrl),
                   ),
-                const SizedBox(height: 8),
-                HtmlWidget(post.content),
-              ],
-            ),
+                ),
+              const SizedBox(height: 8),
+              HtmlWidget(post.content),
+            ],
           ),
         ),
-      );
+      ),
+    );
+  }
 
   static void pushRoute(BuildContext context, _Post post) =>
       Navigator.of(context)
@@ -101,14 +104,14 @@ class _PostScreen extends StatelessWidget {
 class _PostsList extends StatefulWidget {
   final String url;
 
-  const _PostsList(this.url, {Key key}) : super(key: key);
+  const _PostsList(this.url);
 
   @override
   _PostsState createState() => _PostsState();
 }
 
 class _PostsState extends State<_PostsList> {
-  Future<List<_Post>> posts;
+  late final Future<List<_Post>> posts;
 
   @override
   void initState() {
@@ -122,7 +125,7 @@ class _PostsState extends State<_PostsList> {
   @override
   Widget build(BuildContext context) => FutureBuilder<List<_Post>>(
         builder: (context, snapshot) => snapshot.hasData
-            ? _build(snapshot.data)
+            ? _build(snapshot.requireData)
             : snapshot.hasError
                 ? Text(snapshot.error.toString())
                 : const Center(child: CircularProgressIndicator()),
@@ -134,25 +137,28 @@ class _PostsState extends State<_PostsList> {
         itemCount: posts.length,
       );
 
-  Widget _buildItem(_Post post) => ListTile(
-        leading: post.featuredMedia != null
-            ? Image.network(
-                post.featuredMedia.thumbnail,
-                fit: BoxFit.cover,
-                height: 44,
-                width: 44,
-              )
-            : const SizedBox(width: 44),
-        onTap: () => _PostScreen.pushRoute(context, post),
-        subtitle: HtmlWidget(
-          post.excerpt,
-          onTapUrl: (_) {
-            _PostScreen.pushRoute(context, post);
-            return true;
-          },
-        ),
-        title: HtmlWidget(post.title),
-      );
+  Widget _buildItem(_Post post) {
+    final featuredMedia = post.featuredMedia;
+    return ListTile(
+      leading: featuredMedia != null
+          ? Image.network(
+              featuredMedia.thumbnail,
+              fit: BoxFit.cover,
+              height: 44,
+              width: 44,
+            )
+          : const SizedBox(width: 44),
+      onTap: () => _PostScreen.pushRoute(context, post),
+      subtitle: HtmlWidget(
+        post.excerpt,
+        onTapUrl: (_) {
+          _PostScreen.pushRoute(context, post);
+          return true;
+        },
+      ),
+      title: HtmlWidget(post.title),
+    );
+  }
 
   List<_Post> _parseJson(json) {
     final posts = <_Post>[];
@@ -174,36 +180,52 @@ class _PostsState extends State<_PostsList> {
 class _Post {
   final String content;
   final String excerpt;
-  final _Media featuredMedia;
+  final _Media? featuredMedia;
   final int id;
   final String link;
   final String title;
 
   const _Post({
-    this.content,
-    this.excerpt,
+    required this.content,
+    required this.excerpt,
     this.featuredMedia,
-    this.id,
-    this.link,
-    this.title,
+    required this.id,
+    required this.link,
+    required this.title,
   });
 
-  factory _Post.fromJson(Map json) {
-    final embedded = json['_embedded'] as Map;
-    final featuredMediaList = embedded['wp:featuredmedia'];
-    final featuredMedia =
-        featuredMediaList is List ? featuredMediaList.first as Map : null;
+  static _Post? fromJson(Map json) {
+    if (json
+        case {
+          'content': {
+            'rendered': final String content,
+          },
+          'excerpt': {
+            'rendered': final String excerpt,
+          },
+          'id': final int id,
+          'link': final String link,
+          'title': {
+            'rendered': final String title,
+          },
+          '_embedded': final Map embedded,
+        }) {
+      _Media? featuredMedia;
+      if (embedded case {'wp:featuredmedia': [final Map featuredMediaJson]}) {
+        featuredMedia = _Media.fromJson(featuredMediaJson);
+      }
 
-    return _Post(
-      // this is unsafe, do not do this in real app
-      content: (json['content'] as Map)['rendered'] as String,
-      excerpt: (json['excerpt'] as Map)['rendered'] as String,
-      featuredMedia:
-          featuredMedia != null ? _Media.fromJson(featuredMedia) : null,
-      id: json['id'] as int,
-      link: json['link'] as String,
-      title: (json['title'] as Map)['rendered'] as String,
-    );
+      return _Post(
+        content: content,
+        excerpt: excerpt,
+        featuredMedia: featuredMedia,
+        id: id,
+        link: link,
+        title: title,
+      );
+    }
+
+    return null;
   }
 }
 
@@ -215,17 +237,34 @@ class _Media {
   final int width;
 
   const _Media({
-    this.height,
-    this.sourceUrl,
-    this.thumbnail,
-    this.width,
+    required this.height,
+    required this.sourceUrl,
+    required this.thumbnail,
+    required this.width,
   });
 
-  factory _Media.fromJson(Map json) => _Media(
-        height: (json['media_details'] as Map)['height'] as int,
-        sourceUrl: json['source_url'] as String,
-        thumbnail: (((json['media_details'] as Map)['sizes']
-            as Map)['thumbnail'] as Map)['source_url'] as String,
-        width: (json['media_details'] as Map)['width'] as int,
+  static _Media? fromJson(Map json) {
+    if (json
+        case {
+          'media_details': {
+            'height': final int height,
+            'width': final int width,
+            'sizes': {
+              'thumbnail': {
+                'source_url': final String thumbnail,
+              },
+            },
+          },
+          'source_url': final String sourceUrl,
+        }) {
+      return _Media(
+        height: height,
+        sourceUrl: sourceUrl,
+        thumbnail: thumbnail,
+        width: width,
       );
+    }
+
+    return null;
+  }
 }
