@@ -124,7 +124,7 @@ void main() {
         tester,
         html,
         enableCaching: true,
-        baseUrl: Uri.http('domain.com', ''),
+        baseUrl: Uri.http('domain.com'),
       );
       final built2 = helper.buildCurrentState();
       enableCachingExpect(built1, built2, isFalse);
