@@ -23,11 +23,16 @@ void main() {
       _duration = const Duration(milliseconds: 10);
       _playbackEvents = StreamController<PlaybackEventMessage>.broadcast();
 
+      // TODO: remove lint ignore when our minimum Flutter version >= 3.10
+      // ignore: deprecated_member_use
       audioSessionMc.setMockMethodCallHandler((_) async {});
     });
 
     tearDown(() {
-      audioSessionMc.setMockMethodCallHandler(null);
+      // TODO: remove lint ignore when our minimum Flutter version >= 3.10
+      // ignore: deprecated_member_use
+      audioSessionMc.setMockMethodCallHandler((_) async {});
+
       _playbackEvents.close();
     });
 
@@ -82,9 +87,14 @@ void main() {
     });
 
     testWidgets('shows remaining (narrow)', (tester) async {
+      // TODO: remove lint ignore when our minimum Flutter version >= 3.10
+      // ignore: deprecated_member_use
       tester.binding.window.physicalSizeTestValue = const Size(320, 568);
+      // ignore: deprecated_member_use
       addTearDown(tester.binding.window.clearPhysicalSizeTestValue);
+      // ignore: deprecated_member_use
       tester.binding.window.devicePixelRatioTestValue = 1;
+      // ignore: deprecated_member_use
       addTearDown(tester.binding.window.clearDevicePixelRatioTestValue);
 
       _duration = const Duration(minutes: 12, seconds: 34);
